refactor(agent): extract ChunkContextKeyDeriver from ChunkContextFactory
ChunkContextFactory mixed prompt assembly with cache-key derivation.
Split the mapping-fingerprint/cache-key logic into its own stateless
@internal collaborator, mirroring the AuditCommand-collaborator split.
